Move the batching tests to its appropriate namespace/folder
